+++ docs/src/chapters/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+_[0.2.0] - 2025-09-12_
+
+- Adjustments for Penpot 2.9 Typography, upgrade focusring handling, update docs 📑
+
+Old way:
+`border: var(—kfw-focusring-outline);`
+
+New way:
+
+```
+border-width: var(kfw-focusring-outline-width);
+border-style: var(—kfw-focusring-outline-style);
+border-color: var(—kfw-color-fn-active);
+```
+
 _[0.1.3] - 2025-08-22_
 
 - Remove `kfw-color-accent`, use `kfw-color-line-11` for header/footer instead or `kfw-base-color-green-300`.